refactor testsPrepareUtils
