+++ src/test/java/org/opensearch/neuralsearch/sparse/testsPrepareUtils.java
@@ -2,7 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.neuralsearch;
+package org.opensearch.neuralsearch.sparse;
 
 import org.apache.lucene.codecs.Codec;
 import org.apache.lucene.codecs.DocValuesProducer;
